backend/requirements.txt
fastapi
uvicorn
websockets
google-genai
sqlalchemy
aiosqlite
numpy
python-dotenv
pydantic
qdrant-client
sentence-transformers
ollama
tenacity
httpx
